feat: use @polkadot-api/light-client-extension-helpers
