+++ projects/extension/package.json
@@ -43,6 +43,7 @@
     "web-ext": "^7.8.0"
   },
   "dependencies": {
+    "@polkadot-api/light-client-extension-helpers": "pr-164",
     "@substrate/connect-extension-protocol": "*",
     "@unstoppablejs/utils": "^1.1.0",
     "react": "^18.2.0",
